style: 统一文件结尾格式并移除多余空行
- 移除多个文件末尾多余的空行，确保文件以换行符结尾
- 调整部分代码格式，如字段顺序、括号格式等
- 删除不必要的 using 语句
- 统一代码风格，提高可读性

+++ GameFrameX.NetWork.RemoteMessaging/Transport/DefaultMessageCodec.cs
@@ -15,7 +15,6 @@
 //  ==========================================================================================
 
 using System.Buffers.Binary;
-using GameFrameX.NetWork.RemoteMessaging.Contracts;
 using GameFrameX.ProtoBuf.Net;
 
 namespace GameFrameX.NetWork.RemoteMessaging.Transport;
@@ -49,7 +48,7 @@ internal sealed class DefaultMessageCodec : IMessageCodec
         var totalLength = messageData.Length + InnerPackageHeaderLength;
         var buffer = new byte[totalLength];
         BinaryPrimitives.WriteInt32BigEndian(buffer.AsSpan(0, 4), totalLength);
-        buffer[4] = (byte)message.OperationType;
+        buffer[4] = message.OperationType;
         buffer[5] = zipFlag;
         BinaryPrimitives.WriteInt32BigEndian(buffer.AsSpan(6, 4), message.UniqueId);
         BinaryPrimitives.WriteInt32BigEndian(buffer.AsSpan(10, 4), message.MessageId);
